+++ pkg/dds/test/dap/integration/debug_attach_test.dart
@@ -128,7 +128,12 @@ main() {
 
       expect(
         outputEvents.map((e) => e.output).join(),
-        contains('Invalid --vm-service-uri argument: http://bogus.local'),
+        allOf(
+          contains(
+            'Failed to connect/initialize debugger for ws://bogus.local/',
+          ),
+          contains('Failed host lookup'),
+        ),
       );
     });