+++ pkg/linter/lib/src/rules/prefer_is_not_empty.dart
@@ -65,7 +65,7 @@ class _Visitor extends SimpleAstVisitor<void> {
     // Element should also support "isNotEmpty".
     var propertyTarget = propertyElement.enclosingElement2;
     if (propertyTarget == null ||
-        getChildren2(propertyTarget, 'isNotEmpty').isEmpty) {
+        getChildren(propertyTarget, 'isNotEmpty').isEmpty) {
       return;
     }